functions/infrastructure/parser/pdf_parser.py
from application.parser import IParser, ParsingResult
from PyPDF2 import PdfFileReader
from io import BytesIO

class PDFParser(IParser):
    def parse(self, file: BytesIO) -> ParsingResult:
        # Initialize a PDF file reader
        pdf = PdfFileReader(file)
        
        # Initialize the text and images list
        text = ""
        images = []
        
        # Extract text from each page
        for page in range(pdf.getNumPages()):
            text += pdf.getPage(page).extractText()
            
        # TODO: Extract images from the PDF file
        # This part is a bit more complex as it depends on the structure of your PDF file
        # You might need to use a library like PDFMiner or PyMuPDF if PyPDF2 doesn't work
        
        return ParsingResult(text=text, images=tuple(images))
